+++ crates/cli/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## 0.1.2-git
+## 0.2.0-git
+
+### Major
+
+- Move `--serial` and `--timeout` to commands that need them
 
 ### Patch